Revert "Allow fractionnal duration in spnl, for percolation to run less than every seconds."
This reverts commit 70686268b48e8991e28f1b6dd635b29d5f8e12f6.

+++ spnl-core/src/main/scala/com/wajam/spnl/TaskContext.scala
@@ -9,9 +9,9 @@ import com.wajam.spnl.TaskContext._
  * Task running context
  */
 case class TaskContext(var data:Map[String, Any] = Map[String, Any](),
-                       var normalRate: Double = 1,
-                       var throttleRate: Double = 1,
-                       var maxConcurrent: Double = 5) {
+                       var normalRate:Int = 1,
+                       var throttleRate:Int = 1,
+                       var maxConcurrent: Int = 5) {
   def toJson: String = {
     new String(jasonCodec.encode(JObject(List(JField("data", JsonCodec.toJValue(data)))), Encoding))
   }
